// api/src/modules/event/constants/enums/errors.enum.ts
export enum ErrorMessage {
  EVENT_ID_EMPTY_ERROR = 'Bạn vui lòng chọn [sự kiện].',
  LANGUAGE_ID_EMPTY_ERROR = 'Bạn vui lòng chọn [ngôn ngữ].',
  ID_NAN_ERROR = 'Giá trị [id] không hợp lệ.',
  NO_CONTENT = 'Không có dữ liệu hiển thị.',

  OPERATOR_EVENT_ERROR = 'Lưu thông tin [sự kiện] thất bại.',
  EVENT_NOT_FOUND_ERROR = '[Sự kiện] không tồn tại (id: %s)',

  FILE_NOT_FOUND_ERROR = 'Không tìm thấy file.',

  TIME_NAN_ERROR = '[Thời gian] không hợp lệ.',
}